requirements.txt
# ============================================================================
# PXQSDA Project Requirements
# Python 3.8+ dependencies for ML pipeline, gRPC client, and utilities
# ============================================================================

# ============================================================================
# CORE DEPENDENCIES
# ============================================================================
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Data Processing & Visualization
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# ============================================================================
# DISTRIBUTED SYSTEM & COMMUNICATION
# ============================================================================
# gRPC and Protocol Buffers
grpcio>=1.56.0
grpcio-tools>=1.56.0
protobuf>=4.23.0

# ZeroMQ (for verification/monitoring only - C++ core uses native bindings)
pyzmq>=25.0.0

# ============================================================================
# UTILITIES & MONITORING
# ============================================================================
# Progress bars and terminal UI
tqdm>=4.65.0

# Process and system monitoring
psutil>=5.9.0

# Symbolic math (for analysis/diagnostics)
sympy>=1.12

# JSON validation
jsonschema>=4.17.0

# ============================================================================
# LOGGING & CONFIGURATION
# ============================================================================
# Logging
loguru>=0.7.0

# Config management
pyyaml>=6.0

# ============================================================================
# DEVELOPMENT & TESTING (optional)
# ============================================================================
# Uncomment the following for development:
# pytest>=7.3.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.3.0
# ipython>=8.12.0
# jupyter>=1.0.0

# ============================================================================
# INSTALLATION INSTRUCTIONS
# ============================================================================
# Create virtual environment:
#   python3 -m venv ml/venv
#
# Activate virtual environment:
#   source ml/venv/bin/activate  (macOS/Linux)
#   ml\venv\Scripts\activate     (Windows)
#
# Install requirements:
#   pip install --upgrade pip
#   pip install -r requirements.txt
#
# Generate gRPC Python files (if needed):
#   python -m grpc_tools.protoc -I./bridge/pb --python_out=./bridge/pb --grpc_python_out=./bridge/pb ./bridge/pb/prime_bridge.proto
